+++ kafka-online-orders/client-app/src/main/java/dev/sokarau/client/app/ClientApp.java
@@ -1,4 +1,4 @@
-package dev.sokarau.kafkaonlineorders;
+package dev.sokarau.client.app;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,8 +6,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 @SpringBootApplication
 @RestController
-public class KafkaOnlineOrdersApplication {
+public class ClientApp {
 	public static void main(String[] args) {
-		SpringApplication.run(KafkaOnlineOrdersApplication.class, args);
+		SpringApplication.run(ClientApp.class, args);
 	}
 }
