+++ kafka-online-orders/client-app/src/main/java/dev/sokarau/client/app/dto/OrderRepository.java
@@ -1,5 +1,6 @@
-package dev.sokarau.kafkaonlineorders;
+package dev.sokarau.client.app.dto;
 
+import dev.sokarau.client.app.dto.Order;
 import org.bson.types.ObjectId;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
