+++ kafka-online-orders/client-app/src/main/resources/.env.example
