+++ kafka-online-orders/client-app/src/main/resources/application.properties
@@ -1,0 +1,3 @@
+spring.data.mongodb.database=${MONGO_DATABASE}
+spring.data.mongodb.uri=mongodb+srv://${MONGO_USER}:${MONGO_PASSWORD}@${MONGO_CLUSTER}
+spring.kafka.bootstrap-servers=localhost:9092
