--- kafka-online-orders/src/main/resources/application.properties
@@ -1,2 +1,0 @@
-spring.data.mongodb.database=${MONGO_DATABASE}
-spring.data.mongodb.uri=mongodb+srv://${MONGO_USER}:${MONGO_PASSWORD}@${MONGO_CLUSTER}
